New syscall format

+++ include/arcticOS/graphics/primitives.h
@@ -36,7 +36,7 @@ extern int SCREEN_WIDTH;
 extern int SCREEN_HEIGHT;
 
 void graphics_get_screen_size();
-void graphics_plot_pixel(uint16_t x, uint16_t y, uint16_t color);
+void graphics_plot_pixel(int x, int y, uint16_t color);
 void graphics_refresh();
 void graphics_fill(uint16_t color);
 void graphics_rect(uint16_t x, uint16_t y, uint16_t x2, uint16_t y2, uint16_t color);
